add tests to the mapper layer
